#8 [RENAME] Profile 컴포넌트 이름을 Avatar로 변경
