+++ src/components/Avatar/Avatar.jsx
@@ -1,4 +1,4 @@
-import * as S from './Profile.style.jsx';
+import * as S from './Avatar.style.jsx';
 
 export default function Profile({ src, size }) {
   return <S.Profile src={src} $size={size} />;
