+++ src/components/Avatar/index.js
@@ -1,0 +1,1 @@
+export { default as Avatar } from './Avatar';
