--- src/components/Profile/index.js
@@ -1,1 +1,0 @@
-export { default as Profile } from './Profile';
